+++ examples/test-site/src/components/ProductTout/token.tsx
@@ -53,7 +53,6 @@ const asFilterableProductContainer = withDesign({
   ComponentWrapper: asFilterableByGroup(),
 });
 
-export default asProductToutDefaultStyle;
 export {
   asProductToutDefaultStyle,
   asFilterableProductContainer,
