+++ packages/bodiless-components/src/List/withSublist.tsx
@@ -17,7 +17,6 @@ import { flow } from 'lodash';
 import { withToggleButton } from '../Toggle';
 import { ListDesignableComponents, WithSublistToggle } from './types';
 import asBasicSublist from './asBasicSublist';
-import asAccordionSublist from './asAccordionSublist';
 import withDeleteSublistOnUnwrap from './withDeleteSublistOnUnwrap';
 
 /**
@@ -38,11 +37,5 @@ const withBasicSublist = flow(
   withSublist,
 );
 
-const withAccordionSublist = flow(
-  withDeleteSublistOnUnwrap,
-  asAccordionSublist,
-  withSublist,
-);
-
 export default withSublist;
-export { withBasicSublist, withAccordionSublist };
+export { withBasicSublist };
